refactor: move model package to internal
